complete rotation

+++ srcs/isometric.c
@@ -3,10 +3,10 @@
 
 
 
-void	ft_isometric(double *x, double *y, int z)
+void	ft_isometric(double *x, double *y, double z)
 {
-	int	prev_x;
-	int	prev_y;
+	double	prev_x;
+	double	prev_y;
 
 	prev_x = *x;
 	prev_y = *y;
